% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate.model}
\alias{evaluate.model}
\title{Title Simulate conjoint process of profile selection}
\usage{
evaluate.model(input)
}
\arguments{
\item{input}{dataframe with treatment variables generated by generate_sample() and response by simulate_conjoint(), i.e. long conjoint format}
}
\value{
- est_coef: estimated coefficient
}
\description{
Analysis the conjoint process of profile selection and
}
