% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate conjoint process of profile selection}
\usage{
evaluate_model(input, robust = TRUE)
}
\arguments{
\item{input}{dataframe with treatment variables generated by sim_to_long(), i.e. long conjoint format}
}
\value{
- est_coef: estimated coefficient
- est_se: estimated standard error
- true_coef: coefficient from true_coefs
- sig: logical, if the est_coef was significant (alpha = 0.05)
- in_ci95: logical, if the true_coef was in the CI95
- typeS: logical, did the sign of sig. coef match to true?
- typeM: ratio of estimated/true, if significant (for typeS and typeM, see Gelman and Carlin, 2014)
}
\description{
Analysis the conjoint process of profile selection and
}
