% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cj.R
\name{sim_cj}
\alias{sim_cj}
\title{Title}
\usage{
sim_cj(
  n_attributes = NULL,
  n_levels = NULL,
  group_name = NULL,
  units = NULL,
  n_tasks = NULL,
  true_coef = NULL,
  sigma.u_k = NULL,
  n_sim = NULL,
  power = TRUE,
  p
)
}
\arguments{
\item{p}{}
}
\description{
Title
}
