% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_to_long.R
\name{sim_to_long}
\alias{sim_to_long}
\title{Title}
\usage{
sim_to_long(input)
}
\arguments{
\item{input}{dataframe from simulate_conjoint() function}
}
\value{
a long dataframe
}
\description{
Transform the wide conjoint dataframe in a long one for analysis
}
