% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsimPWR-package.R
\docType{package}
\name{cjsimPWR-package}
\alias{cjsimPWR}
\alias{cjsimPWR-package}
\title{cjsimPWR: Power Analyses for Conjoint Experiments Using Simulation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simulation-based power analysis for forced-choice conjoint experiments that estimate average marginal component effects (AMCEs) as defined by Hainmueller, Hopkins and Yamamoto (2014) \doi{10.1093/pan/mpt024}. Choices are simulated from a logit model calibrated so that the true AMCEs match requested effects within a calibration tolerance, optionally with respondent heterogeneity and respondent subgroups. AMCEs and differences between subgroups are estimated by least squares with respondent-clustered standard errors. Power, Type I error and coverage are reported with Monte Carlo standard errors, as are the Type S and Type M errors of Gelman and Carlin (2014) \doi{10.1177/1745691614551642}. Extends the simulation framework of Stefanelli and Lukac (2020) \doi{10.31235/osf.io/spkcy}.
}
\details{
Use \code{\link[=power_sim]{power_sim()}} to simulate repeated paired, forced-choice conjoint experiments and report
power, Type I error, coverage, and Type S and Type M errors with Monte Carlo standard errors.
The default choice model is calibrated to requested average marginal component effects (AMCEs),
optionally with respondent heterogeneity and respondent subgroups. Estimates use
respondent-clustered standard errors.

For individual steps, use \code{\link[=conjoint_design]{conjoint_design()}} to describe the attributes,
\code{\link[=simulate_experiment]{simulate_experiment()}} to draw an experiment, \code{\link[=estimate_amce]{estimate_amce()}} to analyse it,
and \code{\link[=summarise_runs]{summarise_runs()}} to summarise estimates from repeated experiments.
Run \code{news(package = "cjsimPWR")} for release notes and \code{citation("cjsimPWR")} for citations.
}
\seealso{
\code{\link[=power_sim]{power_sim()}}, \code{\link[=conjoint_design]{conjoint_design()}}, \code{\link[=simulate_experiment]{simulate_experiment()}}, \code{\link[=estimate_amce]{estimate_amce()}},
\code{\link[=summarise_runs]{summarise_runs()}}
}
\author{
\strong{Maintainer}: Alberto Stefanelli \email{alberto.stefanelli@yale.edu} [copyright holder]

Authors:
\itemize{
  \item Martin Lukac \email{m.b.lukac@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Andrew G. Timm (Author of the closed-form Type S and Type M code adapted from the 'retrodesign' package) [contributor, copyright holder]
}

}
