% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate the simulated conjoint process of profile selection against the hypothesized true coefficients.}
\usage{
evaluate_model(input, robust = TRUE)
}
\arguments{
\item{input}{list; a list containing (a) a dataframe with the simulated data in long format generated by sim_to_long() and (b) a list of inputs that represents the design specifications.}

\item{robust}{logical; whether the evaluation should be use cluster (id) robust standard error. Default is set to TRUE.}
}
\value{
The function return a dataframe containing the following parameters:
\itemize{
\item{num_respondents:} {int; total number or respondents.}
\item{num_attrbs:} {int; total number of attributes included in the conjoint experiment.}
\item{num_tasks:} {int; total number of task taken by each respondent.}
\item{id_grp:} {character; name of the sub-population of interest.}
\item{attrb:} {int; name of the attribute included in the conjoint design. At the moment these cannot be changed and are in the format var_1 ... var_x.}
\item{level:} {int; levels of each attribute.}
\item{true_coef:} {numeric; coefficient from true_coefs.}
\item{estimate:} {numeric; estimated coefficient from the linear probability model.}
\item{std.error[.robust]:} {numeric; estimated (robust) standard error.}
\item{ci[_robust]_lower:} {numeric; lower bound of the confidence interval calculated (level= 0.95).}
\item{ci[_robust]_upper:} {numeric;upper bound of the confidence interval calculated (level= 0.95).}
\item{in_ci[_robust]:} {logical, if the true_coef is contained in the confidence level.}
\item{sig[_robust]:} {logical, if the estimate coefficent (estimate) is significant (alpha = 0.05).}
\item{typeS[_robust]:} {logical, did the sign of sig. coef match to true? (see, Gelman and Carlin, 2014).}
\item{typeM[_robust]:} {numeric; ratio of estimated/true coefficent, if significant (see Gelman and Carlin, 2014).}
}
}
\description{
Power analysis and error-rate calculations of the conjoint process of profile selection.
}
\examples{
# Conjoint design with 3 attributes with 2, 3, 5 levels respectively
design_example <- generate_design(n_profiles = 2,
 n_attributes = 3,
 n_levels = c(2, 3, 5))

# Design with multiple groups/sub-populations:
# random sample of 500 Democratic, 200 Independent, 500 Republican respondents
sample_subgrp <- generate_samples(design =  list(design_example, design_example, design_example),
                               units = c(500, 200, 500),
                               n_tasks = c(3, 3, 3),
                               group_name = c("Democrat","Independent", "Republican" )
                               )

# Simulated data using the coefficients contained in true_coef for each subgroup of respondents
simulated_cj <- simulate_conjoint(sample_subgrp,
            true_coef = list("Democrat" = list(0.2, c(-0.1, 0.1), c(-0.1, -0.1, -0.1, 0.1)),
                             "Independent" = list(0.1, c(-0.2, -0.05),  c(-0.1, 0.1, 0.1, 0.3)),
                             "Republican" = list(0.1, c(-0.1, 0.05),  c(-0.1, 0.2, -0.1, 0.1))),
           sigma.u_k = 0.05,
           LOG = FALSE)

# Transform the dataframe obtained from simulate_conjoint() in a long format
simulated_cj_long <- sim_to_long(simulated_cj)
# Evaluate the conjoint process of profile selection against the hypothesized true coefficients.
evaluate <- evaluate_model(simulated_cj_long)

}
