% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_design.R
\name{generate_design}
\alias{generate_design}
\title{Generate design}
\usage{
generate_design(n_attributes, n_levels, n_profiles = 2, rem.eq.prof = T)
}
\arguments{
\item{n_attributes}{int; number of independent variables (ie treatments) in the conjoint experiment}

\item{n_levels}{int; vector of ints of length n_attributes; number of levels of variables should be > 1.}

\item{n_profiles}{int; number of profiles to be compared together. always set to 2}

\item{rem.eq.prof}{logical; remove comparisons of same profiles}
}
\value{
dataframe - large dataframe with a full factorial design, which combines all possible levels together.
}
\description{
Generate full factorial experimental design
}
