% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_samples.R
\name{generate_samples}
\alias{generate_samples}
\title{Generate samples}
\usage{
generate_samples(design, units, n_tasks, LOG = FALSE, group_name = NULL)
}
\arguments{
\item{design}{single or a list of dataframe from generate_design()}

\item{units}{int; single or a list indicating the number of respondents}

\item{n_tasks}{int;  single or a list indicating number of tasks performed by each respondent}

\item{LOG}{logical; return debugging information on sample generation}

\item{group_name}{character; optional vector of names for the observed populations/subgroups of respondents}
}
\value{
Samples from the full factorial for each sub populations n_population * units * n_tasks
}
\description{
Sample of profiles from the full factorial design. If multiple (sub)populations are present, generate_sample should be repeated for each population
}
\examples{
Sample of
100 respondents
w 3 tasks for each respondent
sample <- generate_sample(design = design_example,
                                units = 100,
                                n_tasks = 3)

Sample of
500 Democratic
200 Independent
500 Republican
w 3 tasks for each respondent
sample_usa <- generate_sample(design =  list(design_example, design_example, design_example),
                               units = c(500, 200, 500),
                               n_tasks = 3,
                               group_name = c("Democrat","Independent", "Republican" ))


}
