% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_sim.R
\name{power_sim}
\alias{power_sim}
\title{Title}
\usage{
power_sim(
  n_attributes = NULL,
  n_levels = NULL,
  group_name = NULL,
  units = NULL,
  n_tasks = NULL,
  true_coef = NULL,
  sigma.u_k = NULL,
  sim_runs = 100,
  seed = NULL,
  n_sim = NULL,
  p
)
}
\arguments{
\item{p}{}
}
\description{
Title
}
