% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_to_long.R
\name{sim_to_long}
\alias{sim_to_long}
\title{Wide to long transformation of the simulated conjoint data}
\usage{
sim_to_long(input)
}
\arguments{
\item{input}{list; A list containg a dataframe and inputs from simulate_conjoint() function.}
}
\value{
The function returns a list containing:
\itemize{
\item{dataframe:} {tranformed from a wide to a long format}
\item{design specifications:} {the conjoint design specifications (inputs)}
}
}
\description{
Transform the wide conjoint dataframe contained in simulate_conjoint() in a long one for analysis.
}
\examples{
# Conjoint design with 3 attributes with 2, 3, 5 levels respectively
design_example <- generate_design(n_profiles = 2,
 n_attributes = 3,
 n_levels = c(2, 3, 5))

# Design with multiple groups/sub-populations:
# random sample of 500 Democratic, 200 Independent, 500 Republican respondents
sample_subgrp <- generate_samples(design =  list(design_example, design_example, design_example),
                               units = c(500, 200, 500),
                               n_tasks = c(3, 3, 3),
                               group_name = c("Democrat","Independent", "Republican" )
                               )

# Simulated data using the coefficients contained in true_coef for each subgroup of respondents
simulated_cj <- simulate_conjoint(sample_subgrp,
            true_coef = list("Democrat" = list(0.2, c(-0.1, 0.1), c(-0.1, -0.1, -0.1, 0.1)),
                             "Independent" = list(0.1, c(-0.2, -0.05),  c(-0.1, 0.1, 0.1, 0.3)),
                             "Republican" = list(0.1, c(-0.1, 0.05),  c(-0.1, 0.2, -0.1, 0.1))),
           sigma.u_k = 0.05,
           LOG = FALSE)

# Transform the dataframe obtained from simulate_conjoint() in a long format
simulated_cj_long <- sim_to_long(simulated_cj)
}
