% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_conjoint.R
\name{simulate_conjoint}
\alias{simulate_conjoint}
\title{Simulate conjoint process of profile selection.}
\usage{
simulate_conjoint(input, true_coef, sigma.u_k = 0.02, LOG = FALSE)
}
\arguments{
\item{input}{list; List containing (1) a dataframe with the sample and (2) a list of design specifications (input = units, n_tasks, group_name) generated by generate_samples() in a wide conjoint format.}

\item{true_coef}{vector or list of vectors; hypothesized (true) coefficients for all parameters to be estimated. The coefficients must be in order, e.g., var_1_lvl_1, var_1_lvl2, var_2_lvl1, var_2_lvl2.}

\item{sigma.u_k}{standard deviation of the treatment effect for each variable/level for each respondent.}

\item{LOG}{logical; returns parsing of coefficients and sanity checks.}
}
\value{
The function returns a list similar to the one obtained from generate_samples() with the addition of:
\itemize{
\item{column "y":} {a vector of 0s and 1s corresponding to the simulated selection of that respective profile.}
\item{a list of inputs:} {inputs obtained from generate_design(), generate_samples(), and simulate_conjoint() in a wide conjoint format}
}
}
\description{
Simulate conjoint process of profile selection using the object obtained from generate_samples().
}
\examples{
# Conjoint design with 3 attributes with 2, 3, 5 levels respectively
design_example <- generate_design(n_profiles = 2,
 n_attributes = 3,
 n_levels = c(2, 3, 5))

# Design with multiple groups/sub-populations:
# random sample of 500 Democratic, 200 Independent, 500 Republican respondents
sample_subgrp <- generate_samples(design =  list(design_example, design_example, design_example),
                               units = c(500, 200, 500),
                               n_tasks = c(3, 3, 3),
                               group_name = c("Democrat","Independent", "Republican" )
                               )
# Simulated data using the coefficients contained in true_coef for each subgroup of respondents
simulated_cj <- simulate_conjoint(sample_subgrp,
            true_coef = list("Democrat" = list(0.2, c(-0.1, 0.1), c(-0.1, -0.1, -0.1, 0.1)),
                             "Independent" = list(0.1, c(-0.2, -0.05),  c(-0.1, 0.1, 0.1, 0.3)),
                             "Republican" = list(0.1, c(-0.1, 0.05),  c(-0.1, 0.2, -0.1, 0.1))),
           sigma.u_k = 0.05,
           LOG = FALSE)




}
